qt logger now log widget with better format
